Registros tipo valor nomina
Se corrige error sobre el campo valor nomina, cuando se ingresaba un valor de 11 digitos indicaba que no era un valor valido.

+++ java/soi.empresarial.liquidacion/src/com/ach/soi/empresarial/liquidacion/core/LiquidadorActivos.java
@@ -143,7 +143,7 @@ public class LiquidadorActivos {
 			String valorNominaCrudo = bean01.getVlrTotalNomina().getValorCrudo();
 			boolean exceptionNomina = false;
 			try{
-				Integer valorNomina = Integer.valueOf(valorNominaCrudo);
+				Long valorNomina = Long.valueOf(valorNominaCrudo);
 				if ( valorNomina<=0 ){
 					exceptionNomina = true;
 				}
